refactor: move distogram loss benchmarks to proper directory structure
- Move benchmarks to benchmarks/nn/ and benchmarks/nn/functional/
- Follow project convention for benchmark organization
- Separate module and functional benchmarks
